+++ 2019-08-03/erc20-demo/build.rs
@@ -3,6 +3,6 @@ use vergen::{generate_cargo_keys, ConstantsFlags};
 const ERROR_MSG: &str = "Failed to generate metadata files";
 
 fn main() {
-    generate_cargo_keys(ConstantsFlags::all()).expect(ERROR_MSG);
-    println!("cargo:rerun-if-changed=.git/HEAD");
+	generate_cargo_keys(ConstantsFlags::all()).expect(ERROR_MSG);
+	println!("cargo:rerun-if-changed=.git/HEAD");
 }
