+++ 2019-08-03/erc20-demo/runtime/wasm/build.sh
