+++ 2019-08-03/erc20-demo/scripts/init.sh
